+++ src/sound.c
@@ -261,8 +261,8 @@ void sound_load(WAV * wav, const char * name, float min, float max) {
     if (channels > 1) { // convert stereo to mono
         audio = malloc(samplecount * sizeof(short) / 2);
         CHECK_ALLOCATION_ERROR(audio)
-        for (int k = 0; k < samplecount / 2; k++)
-            audio[k] = ((int)samples[k * 2] + (int)samples[k * 2 + 1]) / 2; // prevent overflow
+        for (size_t k = 0; k < samplecount / 2; k++)
+            audio[k] = ((int) samples[k * 2] + (int) samples[k * 2 + 1]) / 2; // prevent overflow
         free(samples);
     }
